finish printf

+++ ft_printf/ft_printf.h
@@ -20,8 +20,8 @@ int		putstrfd(char *str, int fd);
 int		putnbrfd(int nbr, int fd);
 int     putcharfd(int c, int fd);
 int     putpointerfd(void *ptr, int fd);
-int     putunsigned_decimal_fd(unsigned long number, int fd);
-int     puthexfd(unsigned long number,char *base, int fd);
+int     putunsigned_decimal_fd(unsigned int number, int fd);
+int     puthexfd(unsigned int number,char *base, int fd);
 int		ft_printf(const char *input, ...);
 
 #endif
